+++ src/syntax.js
@@ -183,7 +183,9 @@
 
     function renderElementButtons( syntax, syntaxOptions, syntaxLanguage, syntaxButtonsParsed, innerHTMLCopy ) {
         if ( syntaxOptions.showLanguageLabel || syntaxOptions.showCopyButton || syntaxOptions.showPrintButton || syntaxButtonsParsed[ 0 ] ) {
-            var buttons = createElement( "div", "buttons" );
+            var buttons = createElement( "div", "buttons" ),
+                buttonsElements = [];
+
             syntax.appendChild( buttons );
 
             if ( syntaxButtonsParsed[ 0 ] && isDefinedArray( syntaxButtonsParsed[ 1 ] ) ) {
@@ -197,11 +199,14 @@
                         var newCustomButton = createElement( "div", "button" );
                         newCustomButton.innerHTML = customButton.text;
                         newCustomButton.onclick = customButton.onClick;
+                        newCustomButton.style.display = _configuration.buttonsVisible ? "inline-block" : "none";
                         buttons.appendChild( newCustomButton );
 
                         if ( isDefined( customButton.className ) ) {
                             newCustomButton.className += " " + customButton.className;
                         }
+
+                        buttonsElements.push( newCustomButton );
                     }
                 }
             }
@@ -209,6 +214,7 @@
             if ( syntaxOptions.showCopyButton ) {
                 var copyButton = createElement( "div", "button" );
                 copyButton.innerHTML = syntaxOptions.copyButtonText;
+                copyButton.style.display = _configuration.buttonsVisible ? "inline-block" : "none";
                 buttons.appendChild( copyButton );
 
                 copyButton.onclick = function() {
@@ -216,11 +222,14 @@
 
                     fireCustomTrigger( syntaxOptions.onCopy, innerHTMLCopy );
                 };
+
+                buttonsElements.push( copyButton );
             }
 
             if ( syntaxOptions.showPrintButton ) {
                 var printButton = createElement( "div", "button" );
                 printButton.innerHTML = syntaxOptions.printButtonText;
+                printButton.style.display = _configuration.buttonsVisible ? "inline-block" : "none";
                 buttons.appendChild( printButton );
 
                 printButton.onclick = function() {
@@ -253,12 +262,32 @@
 
                     fireCustomTrigger( syntaxOptions.onPrint, newElementForPrint.innerHTML );
                 };
+
+                buttonsElements.push( printButton );
             }
 
             if ( syntaxOptions.showLanguageLabel ) {
                 var languageLabel = createElement( "div", "label" );
                 languageLabel.innerHTML = getFriendlyLanguageName( syntaxLanguage );
                 buttons.appendChild( languageLabel );
+            }
+
+            var buttonsElementsLength = buttonsElements.length;
+
+            if ( buttonsElementsLength > _configuration.maximumButtons ) {
+                var openButton = createElement( "div", "button button-opener" );
+                openButton.innerText = _configuration.buttonsVisible ? _configuration.buttonsCloserText : _configuration.buttonsOpenerText;
+                buttons.insertBefore( openButton, buttons.children[ 0 ] );
+
+                openButton.onclick = function() {
+                    var areButtonsVisible = openButton.innerText === _configuration.buttonsCloserText;
+
+                    for ( var buttonsElementIndex = 0; buttonsElementIndex < buttonsElementsLength; buttonsElementIndex++ ) {
+                        buttonsElements[ buttonsElementIndex ].style.display = areButtonsVisible ? "none" : "inline-block";
+                    }
+
+                    openButton.innerText = areButtonsVisible ? _configuration.buttonsOpenerText : _configuration.buttonsCloserText;
+                };
             }
         }
     }
@@ -639,6 +668,10 @@
         return isDefined( object ) && typeof object === "function";
     }
 
+    function isDefinedNumber( object ) {
+        return isDefined( object ) && typeof object === "number";
+    }
+
     function isDefinedArray( object ) {
         return isDefinedObject( object ) && object instanceof Array;
     }
@@ -717,6 +750,10 @@
 
     function getDefaultArray( value, defaultValue ) {
         return isDefinedArray( value ) ? value : defaultValue;
+    }
+
+    function getDefaultNumber( value, defaultValue ) {
+        return isDefinedNumber( value ) ? value : defaultValue;
     }
 
     function getDefaultStringOrArray( value, defaultValue ) {
@@ -1162,6 +1199,10 @@
     function buildDefaultConfiguration() {
         _configuration.safeMode = getDefaultBoolean( _configuration.safeMode, true );
         _configuration.highlightAllDomElementTypes = getDefaultStringOrArray( _configuration.highlightAllDomElementTypes, [ "div", "code" ] );
+        _configuration.maximumButtons = getDefaultNumber( _configuration.maximumButtons, 2 );
+        _configuration.buttonsVisible = getDefaultBoolean( _configuration.buttonsVisible, true );
+        _configuration.buttonsOpenerText = getDefaultString( _configuration.buttonsOpenerText, "<" );
+        _configuration.buttonsCloserText = getDefaultString( _configuration.buttonsCloserText, ">" );
     }